requirements.txt
django==5.1.4
google-generativeai==0.8.3
python-dotenv==1.0.1
